Refactored the Stream details page into smaller components
